requirements.txt
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
mysql-connector-python==8.1.0
webdriver-manager==4.0.1
python-dotenv==1.0.0 